+++ srcs/utils/vector.c
@@ -6,7 +6,7 @@
 /*   By: vparis <vparis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/04 17:41:23 by vparis            #+#    #+#             */
-/*   Updated: 2019/01/15 19:35:03 by vparis           ###   ########.fr       */
+/*   Updated: 2019/01/28 13:36:09 by vparis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,8 +18,8 @@ t_vec2		vec_rotate(t_vec2 dir, t_float angle)
 	t_float old_x;
 
 	old_x = dir.x;
-	dir.x = dir.x * cos(-angle) - dir.y * sin(-angle);
-	dir.y = old_x * sin(-angle) + dir.y * cos(-angle);
+	dir.x = dir.x * cosf(-angle) - dir.y * sinf(-angle);
+	dir.y = old_x * sinf(-angle) + dir.y * cosf(-angle);
 	return (dir);
 }
 
@@ -33,7 +33,7 @@ t_float		vec_dot(t_vec2 v1, t_vec2 v2)
 
 t_float		vec_len(t_vec2 v1)
 {
-	return (sqrt(vec_dot(v1, v1)));
+	return (sqrtf(vec_dot(v1, v1)));
 }
 
 t_vec2		vec_norm(t_vec2 v1)
@@ -41,9 +41,9 @@ t_vec2		vec_norm(t_vec2 v1)
 	t_float	len;
 
 	len = vec_len(v1);
-	if (len > 0.0)
+	if (len > 0.0f)
 	{
-		len = 1.0 / len;
+		len = 1.0f / len;
 		v1 *= len;
 	}
 	return (v1);
@@ -53,6 +53,6 @@ t_vec2		vec_rotate_vec(t_vec2 v1, t_vec2 v2)
 {
 	t_float	angle;
 
-	angle = acos(vec_dot(v1, v2));
+	angle = acosf(vec_dot(v1, v2));
 	return (vec_rotate(v1, angle));
 }
